+++ src/Silex/Provider/TwigServiceProvider.php
@@ -11,8 +11,7 @@
 
 namespace Silex\Provider;
 
-use Silex\Application;
-use Silex\ServiceProviderInterface;
+use Silex\Api\ServiceProviderInterface;
 use Symfony\Bridge\Twig\Extension\RoutingExtension;
 use Symfony\Bridge\Twig\Extension\TranslationExtension;
 use Symfony\Bridge\Twig\Extension\FormExtension;
@@ -28,7 +27,7 @@ use Symfony\Bridge\Twig\Form\TwigRenderer;
  */
 class TwigServiceProvider implements ServiceProviderInterface
 {
-    public function register(Application $app)
+    public function register(\Pimple $app)
     {
         $app['twig.options'] = array();
         $app['twig.form.templates'] = array('form_div_layout.html.twig');
@@ -38,16 +37,16 @@ class TwigServiceProvider implements ServiceProviderInterface
         $app['twig'] = $app->share(function ($app) {
             $app['twig.options'] = array_replace(
                 array(
-                    'charset'          => $app['charset'],
-                    'debug'            => $app['debug'],
-                    'strict_variables' => $app['debug'],
+                    'charset'          => isset($app['charset']) ? $app['charset'] : 'UTF-8',
+                    'debug'            => isset($app['debug']) ? $app['debug'] : false,
+                    'strict_variables' => isset($app['debug']) ? $app['debug'] : false,
                 ), $app['twig.options']
             );
 
             $twig = new \Twig_Environment($app['twig.loader'], $app['twig.options']);
             $twig->addGlobal('app', $app);
 
-            if ($app['debug']) {
+            if (isset($app['debug']) && $app['debug']) {
                 $twig->addExtension(new \Twig_Extension_Debug());
             }
 
@@ -68,9 +67,6 @@ class TwigServiceProvider implements ServiceProviderInterface
                     $app['fragment.renderer.hinclude']->setTemplating($twig);
 
                     $twig->addExtension(new HttpKernelExtension($app['fragment.handler']));
-                } else {
-                    // fallback for BC, to be removed in 1.3
-                    $twig->addExtension(new TwigCoreExtension());
                 }
 
                 if (isset($app['form.factory'])) {
@@ -108,9 +104,5 @@ class TwigServiceProvider implements ServiceProviderInterface
                 $app['twig.loader.filesystem'],
             ));
         });
-    }
-
-    public function boot(Application $app)
-    {
     }
 }